Rename `trade_shuffle` to `ShuffleTradeItems`
