+++ soh/src/overlays/actors/ovl_En_Ds/z_en_ds.c
@@ -6,7 +6,6 @@
 
 #include "z_en_ds.h"
 #include "objects/object_ds/object_ds.h"
-#include "soh/Enhancements/randomizer/trade_shuffle.h"
 #include "soh/OTRGlobals.h"
 #include "soh/ResourceManagerHelpers.h"
 #include "soh/Enhancements/game-interactor/GameInteractor_Hooks.h"
